// frontend/src/generated/models/documentInfo.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AI Enablement Platform
 * Enterprise AI enablement analysis and recommendations platform
 * OpenAPI spec version: 1.0.0
 */
import type { DocumentStatus } from './documentStatus';
import type { DocumentInfoQualityScore } from './documentInfoQualityScore';
import type { ChunkingStrategy } from './chunkingStrategy';
import type { DocumentInfoCompanyHrDataset } from './documentInfoCompanyHrDataset';
import type { DocumentInfoProcessingCompletedAt } from './documentInfoProcessingCompletedAt';
import type { DocumentInfoDocumentMetadata } from './documentInfoDocumentMetadata';

/**
 * Document information response
 */
export interface DocumentInfo {
  id: number;
  filename: string;
  original_filename: string;
  file_size: number;
  mime_type: string;
  status: DocumentStatus;
  total_chunks: number;
  duplicate_chunks_count?: number;
  total_characters: number;
  quality_score: DocumentInfoQualityScore;
  chunking_strategy: ChunkingStrategy;
  qa_rag_enabled: boolean;
  qa_pairs_generated: number;
  customer_id: string;
  company_hr_dataset?: DocumentInfoCompanyHrDataset;
  created_at: string;
  processing_completed_at: DocumentInfoProcessingCompletedAt;
  document_metadata: DocumentInfoDocumentMetadata;
}